Lib: Ensure that every task that relies on the CWD sets it first
Before, tasks implicitly relied on the CWD being correct. This actually
broke the _pushBranch task, which assumed to run in the repo directory.
Since the CDN related tasks change the CWD, that wasn't correct anymore.

In the future, all tasks that depend on the correct CWD need to
expliclty set it first.

Fixes #14